requirements.txt
accelerate==0.18.0
addict==2.4.0
apptools==5.2.0
backports-functools-lru-cache==1.6.4
black==22.6.0
calibur==0.0.1
cattrs==23.1.2
chumpy==0.70
click==8.1.3
clip
coloredlogs==15.0.1
configargparse==1.5.3
configobj==5.0.6
constants==0.6.0
contourpy==1.0.6
cycler==0.11.0
cython==0.29.32
dash==2.7.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
einops==0.7.0
envisage==6.1.0
exceptiongroup==1.1.3
filelock==3.13.1
flake8==5.0.4
flake8-bugbear==22.8.22
flake8-comprehensions==3.10.0
flask==2.2.2
flatbuffers==23.5.26
fonttools==4.36.0
freetype-py==2.3.0
fsspec==2024.2.0
ftfy==6.1.1
h5py==3.8.0
huggingface-hub==0.20.3
humanfriendly==10.0
imageio==2.21.1
imageio-ffmpeg==0.4.8
importlib-metadata==5.0.0
ipdb==0.13.13
ipydatawidgets==4.3.2
ipympl==0.9.3
itsdangerous==2.1.2
joblib==1.1.0
kiwisolver==1.4.4
kornia==0.6.11
lap==0.4.0
lapsolver==1.1.0
libcst==0.4.7
markdown-it-py==3.0.0
matplotlib==3.7.5
mayavi==4.8.0
mccabe==0.7.0
mdurl==0.1.2
meshio==5.3.4
meshplex==0.19.1
moreorless==0.4.0
mpmath==1.2.1
mypy-extensions==0.4.3
natsort==8.4.0
networkx==2.8.6
ninja==1.11.1
npx==0.1.1
ntplib==0.4.0
numpy==1.23.2
nvdiffrast==0.3.0
onnx==1.15.0
onnxruntime==1.17.1
open-clip-torch==2.24.0
open3d==0.11.2
opencv-python==4.6.0.66
pandas==1.4.3
pathspec==0.9.0
pip==24.0
platformdirs==2.5.2
plotly==5.10.0
plyfile==0.7.4
progress==1.4
protobuf==4.25.3
psutil==5.9.5
py-machineid==0.4.6
pycocotools==2.0.7
pycodestyle==2.9.1
pyface==7.4.2
pyflakes==2.5.0
pyglet==1.5.27
pygments==2.16.1
pynacl==1.5.0
pyntcloud==0.3.1
pyopengl==3.1.0
pyquaternion==0.9.9
pyrender==0.1.45
python-license-manager==0.2.11
pythreejs==2.4.1
pywavelets==1.3.0
regex==2022.10.31
requests-cache==1.1.0
rich==13.6.0
rich-argparse==1.4.0
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.0
seaborn==0.11.2
sentencepiece==0.2.0
simple-romp==1.0.5
sklearn==0.0.post1
smplx==0.1.28
stdlibs==2022.6.8
sympy==1.4
tenacity==8.0.1
tensorboardx==2.6.2.2
threadpoolctl==3.1.0
tifffile==2022.8.12
timm==0.6.7
toml==0.10.2
tomli==2.0.1
tomli-w==1.0.0
torchtyping==0.1.4
trailrunner==1.2.1
traits==6.4.1
traitsui==7.4.0
traittypes==0.2.1
transforms3d==0.4.1
trimesh==3.9.35
typeguard==3.0.2
typing-extensions==4.5.0
typing-inspect==0.8.0
url-normalize==1.4.3
usort==1.0.4
vtk==9.0.1
werkzeug==2.2.2
x21==0.4.25
learnable-primitives